+++ rocks/resolve.py
@@ -3,7 +3,6 @@
 import asyncio
 import pickle
 import re
-import warnings
 import sys
 
 import aiohttp
@@ -15,6 +14,7 @@ from rich.progress import Progress
 
 from rocks import cli
 from rocks import index
+from rocks import logging
 
 # Run asyncio nested for jupyter notebooks, GUIs, ...
 nest_asyncio.apply()
@@ -58,7 +58,7 @@ def identify(id_, return_id=False, local=True, progress=False):
     elif isinstance(id_, (set, range)):
         id_ = list(id_)
     elif id_ is None:
-        warnings.warn(f"Received id_ of type {type(id_)}.")
+        logging.logger.warning(f"Received id_ of type {type(id_)}.")
         return (None, np.nan) if not return_id else (None, np.nan, None)  # type: ignore
     elif not isinstance(id_, (list, np.ndarray)):
         raise TypeError(
@@ -67,7 +67,7 @@ def identify(id_, return_id=False, local=True, progress=False):
         )
 
     if not id_:
-        warnings.warn("Received empty list of identifiers.")
+        logging.logger.warning("Received empty list of identifiers.")
         return (None, np.nan) if not return_id else (None, np.nan, None)  # type: ignore
 
     # ------
@@ -132,10 +132,10 @@ async def _identify(id_, local, progress_bar, task):
 async def _resolve(id_, session, local, progress_bar, task):
     """Resolve the identifier locally or remotely."""
 
-    # if pd.isnull(id_) or not id_:  # covers None, np.nan, empty string
-    #     warnings.warn("Received empty or NaN identifier.")
-    #     progress_bar.update(task, advance=1)
-    #     return (None, np.nan, None)
+    if np.nan(id_) or not id_ or id is None:
+        logging.logger.warning("Received empty or NaN identifier.")
+        progress_bar.update(task, advance=1)
+        return (None, np.nan, None)
 
     if local:
         success, (name, number, ssodnet_id) = _local_lookup(id_)
@@ -281,7 +281,7 @@ def _standardize_id_for_quaero(id_):
         else:
             pass
     else:
-        warnings.warn(
+        logging.logger.warning(
             f"Did not understand type of id: {type(id_)}"
             "\nShould be integer, float, or string."
         )
@@ -321,11 +321,11 @@ async def _query_quaero(id_, session):
         return None
 
     if "data" not in response.keys():  # no match found
-        warnings.warn(f"Could not find data for id {id_}.")
+        logging.logger.warning(f"Could not find match for id {id_}.")
         return False
 
     elif not response["data"]:  # empty response
-        warnings.warn(f"Could not find match for id {id_}.")
+        logging.logger.warning(f"Could not find match for id {id_}.")
         return False
 
     return response
@@ -360,7 +360,7 @@ def _parse_quaero_response(data, id_):
             break
     else:
         # Unclear which match is correct.
-        warnings.warn(f"Could not find match for id {id_}.")
+        logging.logger.warning(f"Could not find match for id {id_}.")
         return (None, np.nan, None)
 
     # Found match