+++ rocks/ssodnet.py
@@ -3,7 +3,6 @@
 
 import asyncio
 from itertools import product
-import warnings
 
 import aiohttp
 import json
@@ -12,6 +11,7 @@ import pandas as pd
 from rich.progress import Progress
 
 from rocks import config
+from rocks import logging
 
 
 def get_ssocard(id_ssodnet, progress=False, local=True):
@@ -44,7 +44,7 @@ def get_ssocard(id_ssodnet, progress=False, local=True):
     elif isinstance(id_ssodnet, (set, tuple)):
         id_ssodnet = list(id_ssodnet)
     elif id_ssodnet is None:
-        warnings.warn(f"Received SsODNet ID of type {type(id_ssodnet)}.")
+        logging.logger.warning(f"Received SsODNet ID of type {type(id_ssodnet)}.")
         return [(None, np.nan, None)]
     elif not isinstance(id_ssodnet, (list, np.ndarray)):
         raise TypeError(
@@ -129,13 +129,15 @@ async def _query_ssodnet(id_ssodnet, session):
     response = await session.request(method="GET", url=URL)
 
     if not response.ok:
-        warnings.warn(f"ssoCard query failed for ID: {id_ssodnet}")
+        logger.logger.warning(f"ssoCard query failed for ID: {id_ssodnet}")
         return None
 
     response_json = await response.json()
 
     if response_json is None:
-        warnings.warn(f"ssoCard query returned empty ssoCard for ID: {id_ssodnet}")
+        logging.logger.warning(
+            f"ssoCard query returned empty ssoCard for ID: {id_ssodnet}"
+        )
         return None
 
     return response_json
@@ -231,7 +233,7 @@ def get_datacloud_catalogue(id_ssodnet, catalogue, progress=False, local=True):
     elif isinstance(id_ssodnet, (set, tuple)):
         id_ssodnet = list(id_ssodnet)
     elif id_ssodnet is None:
-        warnings.warn(f"Received SsODNet ID of type {type(id_ssodnet)}.")
+        logging.logger.warning(f"Received SsODNet ID of type {type(id_ssodnet)}.")
         return [(None, np.nan, None)]
     elif not isinstance(id_ssodnet, (list, np.ndarray)):
         raise TypeError(
@@ -360,7 +362,9 @@ async def _query_datacloud(id_ssodnet, catalogue, session):
     response = await session.request(method="GET", url=URL)
 
     if not response.ok:
-        warnings.warn(f"Catalogue query failed for ID: {id_ssodnet} - {catalogue}")
+        logging.logger.warning(
+            f"Catalogue query failed for ID: {id_ssodnet} - {catalogue}"
+        )
         return {"data": {id_ssodnet: {"datacloud": None}}}
 
     response_json = await response.json()